# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restful E-Commerce",
    "version": "1.0.0",
    "description": "A simple Node E-Commerce application for testing RESTful web services"
  },
  "servers": [
    {
      "url": "http://localhost:3004"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Order": {
        "type": "object",
        "required": [
          "user_id",
          "product_id",
          "product_name",
          "product_amount",
          "qty",
          "tax_amt",
          "total_amt"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "description": "The ID of the user"
          },
          "product_id": {
            "type": "string",
            "description": "The ID of the product"
          },
          "product_name": {
            "type": "string",
            "description": "The name of the product"
          },
          "product_amount": {
            "type": "number",
            "description": "The price of the product"
          },
          "qty": {
            "type": "integer",
            "description": "The quantity of the product"
          },
          "tax_amt": {
            "type": "number",
            "description": "The tax amount for the order"
          },
          "total_amt": {
            "type": "number",
            "description": "The total amount for the order"
          }
        }
      },
      "AddOrderResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Orders added successfully!"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "GetAllOrdersResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Orders fetched successfully!"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "GetOrderResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Order found!!"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "UpdateOrderResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Order updated successfully!"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
      "PartialUpdateOrderResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Order updated successfully!"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Authentication successful!"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Up and Running"
          },
          "uptime": {
            "type": "number",
            "description": "Server uptime in seconds",
            "example": 123456
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00.000Z"
          }
        }
      },
      "ImageUploadResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "File uploaded successfully"
          },
          "file": {
            "type": "object",
            "properties": {
              "originalName": {
                "type": "string",
                "description": "Original name of the uploaded file",
                "example": "example.png"
              },
              "path": {
                "type": "string",
                "description": "Server path or URL where the file is stored",
                "example": "uploads/example_123456.png"
              },
              "size": {
                "type": "number",
                "description": "Size of the uploaded file in bytes",
                "example": 45321
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/addOrder": {
      "post": {
        "summary": "Add a list of orders",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Orders added successfully!, Added orders are returned in response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request Payload must be an array of orders! \n\n Each order must have user_id, product_id, product_name, product_amount, qty, tax_amt, and total_amt!"
          }
        }
      }
    },
    "/getAllOrders": {
      "get": {
        "summary": "Get all orders",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "Orders fetched successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllOrdersResponse"
                }
              }
            }
          },
          "404": {
            "description": "No order found!!"
          }
        }
      }
    },
    "/getOrder": {
      "get": {
        "summary": "Get order by various parameters",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "The order ID"
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user"
          },
          {
            "in": "query",
            "name": "product_id",
            "schema": {
              "type": "string"
            },
            "description": "The ID of the product"
          }
        ],
        "responses": {
          "200": {
            "description": "Order found!!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "No order found with the given parameters!"
          }
        }
      }
    },
    "/updateOrder/{id}": {
      "put": {
        "summary": "Update an order by ID",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the order"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate token! \n\n Each Order must have user_id, product_id, product_name, product_amount, qty, tax_amt, and total_amt!"
          },
          "403": {
            "description": "Forbidden! Token is missing!"
          },
          "404": {
            "description": "No order found with the given Id!"
          }
        }
      }
    },
    "/partialUpdateOrder/{id}": {
      "patch": {
        "summary": "Partially update an order by ID",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the order"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "A partial object containing the fields to update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully, updated order is returned in response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartialUpdateOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate token! \n\n Invalid request, no data provided to update!"
          },
          "403": {
            "description": "Forbidden! Token is missing!"
          },
          "404": {
            "description": "No Order found with the given Order Id!"
          }
        }
      }
    },
    "/deleteOrder/{id}": {
      "delete": {
        "summary": "Delete an order by ID",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the order"
          }
        ],
        "responses": {
          "204": {
            "description": "No content is returned in the response."
          },
          "400": {
            "description": "Failed to authenticate token!"
          },
          "403": {
            "description": "Forbidden! Token is missing!"
          },
          "404": {
            "description": "No Order found with the given Order Id!!"
          }
        }
      }
    },
    "/deleteAllOrders": {
      "delete": {
        "summary": "Deletes all the orders in the system",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No content is returned in the response."
          },
          "400": {
            "description": "Failed to authenticate token!"
          },
          "403": {
            "description": "Forbidden! Token is missing!"
          },
          "404": {
            "description": "No orders available to delete!"
          }
        }
      }
    },
    "/auth": {
      "post": {
        "summary": "Authenticate a user and return a JWT",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authentication successful!, token returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Username and Password is required for authentication!"
          },
          "401": {
            "description": "Authentication Failed! Invalid username or password!"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check of the server",
        "responses": {
          "200": {
            "description": "Server is running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server is down"
          }
        }
      }
    },
    "/imageUpload": {
      "post": {
        "summary": "Upload an image file",
        "description": "Upload a JPEG, JPG, or PNG file (up to 5MB) with Bearer token authentication.",
        "tags": [
          "Image Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file to upload (JPG, JPEG, PNG only, max size 5MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – Invalid file format or file size exceeds 5MB"
          },
          "401": {
            "description": "Unauthorized – Missing or invalid Bearer token"
          },
          "403": {
            "description": "Forbidden – Invalid or expired token"
          }
        }
      }
    }
  },
  "tags": []
}